+++ Sources/MenuUSUM/USUM.hpp
@@ -6,9 +6,7 @@
 #include "Includes/USUM4.hpp"
 #include "Includes/USUM5.hpp"
 
-#include "../Helpers.hpp"
-#include "../Helpers/PK67.hpp"
-#include "../Helpers/HotkeyHelpers.hpp"
+#include "../Helpers/Includes/Helpers.hpp"
 
 using namespace CTRPluginFramework;
