+++ Sources/MenuXY/Sources/XY1.cpp
@@ -1,7 +1,6 @@
 #include <CTRPluginFramework.hpp>
 
-#include "../../Helpers/KeyboardHelpers.hpp"
-#include "../../Helpers.hpp"
+#include "../../Helpers/Includes/Helpers.hpp"
 
 using namespace CTRPluginFramework;
